requirements.txt
tqdm~=4.66.1
numpy~=1.26.2
torch~=2.2.0
numba~=0.58.1
paretoset~=1.2.3
scikit-learn~=1.3.2
setuptools~=65.5.0
SQLAlchemy~=2.0.29
python-dotenv~=1.0.1
transformers~=4.40.2
soundfile~=0.12.1
librosa~=0.10.2
Flask~=3.0.3
lightning~=2.2.4
scipy~=1.11.3